+++ semana12/daniel-labeddit/README.md
@@ -1,70 +1,44 @@
-# Getting Started with Create React App
+<div Align='center'>
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+  # Labeddit 
+  <p Align="center">link do surge:   <a href="https://daniel-labeddit.surge.sh/"> Labeddit</a></p>
+  <p Align="center">Projeto com proposta de imitar as funcionalidades básicas do Reddit utilizando a API criada pela Labenu</p> 
 
-## Available Scripts
+</br>
 
-In the project directory, you can run:
 
-### `npm start`
+### Features
 
-Runs the app in the development mode.\
-Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
+- [x] Login/Logout
+- [x] Autorização de acesso às páginas
+- [x] Criar Postagens ou Comentários
+- [x] VotePost / CommentVote
 
-The page will reload if you make edits.\
-You will also see any lint errors in the console.
+</br>
+</br>
 
-### `npm test`
+## 🛠 Tecnologias
 
-Launches the test runner in the interactive watch mode.\
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
+As seguintes ferramentas foram usadas na construção do projeto:
 
-### `npm run build`
+- [React](https://pt-br.reactjs.org/)
+- [Javascript](https://developer.mozilla.org/pt-BR/docs/Web/JavaScript)
+- [CSS](https://developer.mozilla.org/pt-BR/docs/Web/CSS)
+- [API REST](https://documenter.getpostman.com/view/7549981/SW12yx56?version=latest)
 
-Builds the app for production to the `build` folder.\
-It correctly bundles React in production mode and optimizes the build for the best performance.
+</br>
+</br>
 
-The build is minified and the filenames include the hashes.\
-Your app is ready to be deployed!
+### Autor
+---
 
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+<a href="https://www.linkedin.com/in/daniel-ueno-58a8631b4/">
+ <img style="border-radius: 50%;" src="https://avatars.githubusercontent.com/u/81266120?v=4" width="100px;" alt=""/>
+ <br />
+ <sub><b>Daniel Ueno</b></sub></a> <a href="https://www.linkedin.com/in/daniel-ueno-58a8631b4/" title="Rocketseat">🚀</a>
 
-### `npm run eject`
 
-**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
+Feito por Daniel Ueno - Entre em contato!
 
-If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
-
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
-
-You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
-
-## Learn More
-
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
-
-### Code Splitting
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/code-splitting](https://facebook.github.io/create-react-app/docs/code-splitting)
-
-### Analyzing the Bundle Size
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size)
-
-### Making a Progressive Web App
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app](https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app)
-
-### Advanced Configuration
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/advanced-configuration](https://facebook.github.io/create-react-app/docs/advanced-configuration)
-
-### Deployment
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/deployment](https://facebook.github.io/create-react-app/docs/deployment)
-
-### `npm run build` fails to minify
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
+[![Linkedin Badge](https://img.shields.io/badge/-Daniel-blue?style=flat-square&logo=Linkedin&logoColor=white&link=https://www.linkedin.com/in/daniel-ueno/)](https://www.linkedin.com/in/daniel-ueno-58a8631b4/) 
+[![Gmail Badge](https://img.shields.io/badge/-ueno2005@gmail.com-c14438?style=flat-square&logo=Gmail&logoColor=white&link=mailto:ueno2005@gmail.com)](mailto:ueno2005@gmail.com)
