+++ semana12/daniel-labeddit/src/global/GlobalState.js
@@ -1,22 +1,13 @@
 import { GlobalStateContext } from "./GlobalStateContext";
 import React, { useContext, useEffect, useState } from "react";
-import { useRequestData } from "../hooks/useRequestData";
-import { BASE_URL } from "../constants/urls";
 
 export const GlobalState = (props) => {
   const [data, setData] = useState([]);
-
   const token = localStorage.getItem("token");
   const [rightButtonText, setRightButtonText] = useState(
     token ? "Logout" : "Login"
   );
-  const posts = useRequestData(`${BASE_URL}/posts`, {
-    headers: { Authorization: localStorage.getItem("token") },
-  }).data;
 
-  useEffect(() => {
-    setData(posts);
-  }, [posts]);
   const states = { data, rightButtonText };
   const setters = { setData, setRightButtonText };
   return (
